gemini_backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
agno>=1.5.3
google-genai>=1.2.0
mcp>=1.9.1
httpx-sse
starlette
pinecone-client>=3.0.0
